% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin451app-package.R
\docType{package}
\name{fin451app-package}
\alias{fin451app}
\alias{fin451app-package}
\title{fin451app: Commodity Futures Risk & Analytics Dashboard}
\description{
A Golem Shiny app for analyzing commodity futures market dynamics,
volatility, co-dynamics, seasonality, and hedge ratios.
}
\author{
\strong{Maintainer}: Student Name \email{student@ualberta.ca}

}
