% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{cache_app_data}
\alias{cache_app_data}
\title{Cache application data to inst/data-cache/app_data.rds}
\usage{
cache_app_data()
}
\description{
Run this once locally (with FRED_API_KEY set) to pre-populate the cache
so the Docker image can serve data without hitting APIs at runtime.
}
